smpRate and samplesPerSec attributes recognized
SMVSettings has one attribute smpRate and one sub-element SmpRate. The
first is handled by EMF as an attribute smpRateAttribute with metadata
giving its true name (smpRate). However, if the kind=attribute metadata
is not present, the name metadata is not taken into acccount. And if the
name, kind and namespace metadata are given, it does not work either !