review and correct getNamespace() implementations
A preliminary specification about how to find the namespace of a given
element has been written, it is under review. This commit implements
most of this specification (SDI and SDO still missing) to verify that it
can be done. Results of tests on models have not changed after this
commit.