give access to content of Private element
A derived attribute, value, has been added to get the content as text of
a Private element. It was already available by the mixed attribute, but
there is a prefix (and a postfix "]") in the value of this attribute to
indicate that the content is text. The value attribute remove these.